add change admin role

+++ src/main/java/com/borman/geneobook/controllers/AdminController.java
@@ -1,34 +1,45 @@
 package com.borman.geneobook.controllers;
 
+import com.borman.geneobook.entity.Role;
 import com.borman.geneobook.entity.User;
+import com.borman.geneobook.service.RoleService;
 import com.borman.geneobook.service.UserServiceImpl;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.security.Principal;
+import java.util.HashSet;
+import java.util.Objects;
+import java.util.Set;
+
 @Controller
+@RequestMapping("/admin")
 public class AdminController {
 
     private final UserServiceImpl userService;
+    private final RoleService roleService;
 
-    public AdminController(UserServiceImpl userService) {
+    public AdminController(UserServiceImpl userService, RoleService roleService) {
         this.userService = userService;
+        this.roleService = roleService;
     }
 
-    @GetMapping("/admin")
+    @GetMapping("")
     public String userAdmin() {
         return "admin/admin";
     }
 
-    @GetMapping("/admin/listUsers")
+    @GetMapping("/listUsers")
     public String usersList(Model model) {
         model.addAttribute("allUsers", userService.allUsers());
         return "admin/list-users";
     }
 
-    @GetMapping("/admin/listUsers/blocked")
+    @GetMapping("/listUsers/blocked")
     public String usersBlockedById(@RequestParam Long id,  @RequestParam int value) {
         User findUser = userService.findByUserId(id);
         findUser.setEnabled(value);
@@ -36,18 +47,52 @@ public class AdminController {
         return "redirect:/admin/listUsers";
     }
 
-    @GetMapping("/admin/listUsers/delete")
+    @GetMapping("/listUsers/delete")
     public String  deleteUserByIdQuestion(@RequestParam Long id, Model model) {
         User user = userService.findByUserId(id);
         model.addAttribute("deleteUser", user);
         return "admin/delete-user-confirmation";
     }
 
-    @PostMapping("/admin/listUsers/delete")
-    public String  deleteUserByIdSubmit(Model model, @RequestParam String action, User deleteUser) {
+    @PostMapping("/listUsers/delete")
+    public String  deleteUserByIdSubmit(@RequestParam String action, User deleteUser, Principal principal) {
+
+        if (Objects.equals(userService.findByUserName(principal.getName()).getId(), deleteUser.getId())) {
+            System.out.println("Attempting to delete your account");
+            return "redirect:/admin/listUsers";
+        }
+
         if ("okButton".equals(action)) {
             userService.deleteUserById(deleteUser.getId());
         }
+
         return "redirect:/admin/listUsers";
     }
+
+    @GetMapping("/listUsers/change-admin-role")
+    public String addAdminRole(@RequestParam Long id, Principal principal) {
+
+        User user = userService.findByUserName(principal.getName());
+
+        if (Objects.equals(user.getId(), id)) {
+            System.out.println("Trying to edit your rights");
+            return "redirect:/admin/listUsers";
+        }
+
+        Set<Role> roles = user.getRoleSet();
+
+        if (userService.hasRoleAdmin(id)) {
+            roles.remove(roleService.getAdminRole());
+        } else {
+            roles.add(roleService.getAdminRole());
+        }
+
+        user.setRoleSet(roles);
+
+        userService.saveUser(user);
+
+        return "redirect:/admin/listUsers";
+    }
+
+
 }
